requirements.txt
Flask==3.0.2
boto3==1.34.51
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
click==8.1.7